+++ stdlib/LinearAlgebra/src/bidiag.jl
@@ -198,8 +198,8 @@ function svd!(M::Bidiagonal{<:BlasReal}; full::Bool = false)
     d, e, U, Vt, Q, iQ = LAPACK.bdsdc!(M.uplo, 'I', M.dv, M.ev)
     SVD(U, d, Vt)
 end
-function svd(M::Bidiagonal; full::Bool = false)
-    svd!(copy(M), full = full)
+function svd(M::Bidiagonal; kw...)
+    svd!(copy(M), kw...)
 end
 
 ####################