+++ stdlib/LinearAlgebra/src/svd.jl
@@ -54,6 +54,12 @@ function SVD{T}(U::AbstractArray, S::AbstractVector{Tr}, Vt::AbstractArray) wher
         convert(AbstractArray{T}, Vt))
 end
 
+
+abstract type SVDAlgorithm end
+struct DivideAndConquer <: SVDAlgorithm end
+struct Simple <: SVDAlgorithm end
+
+
 # iteration for destructuring into components
 Base.iterate(S::SVD) = (S.U, Val(:S))
 Base.iterate(S::SVD, ::Val{:S}) = (S.S, Val(:V))
@@ -61,7 +67,7 @@ Base.iterate(S::SVD, ::Val{:V}) = (S.V, Val(:done))
 Base.iterate(S::SVD, ::Val{:done}) = nothing
 
 """
-    svd!(A; full::Bool = false) -> SVD
+    svd!(A; full::Bool = false, alg::SVDAlgorithm = DivideAndConquer()) -> SVD
 
 `svd!` is the same as [`svd`](@ref), but saves space by
 overwriting the input `A`, instead of creating a copy.
@@ -92,18 +98,25 @@ julia> A
   0.0       0.0  -2.0  0.0  0.0
 ```
 """
-function svd!(A::StridedMatrix{T}; full::Bool = false) where T<:BlasFloat
+function svd!(A::StridedMatrix{T}; full::Bool = false, alg::SVDAlgorithm = DivideAndConquer()) where T<:BlasFloat
     m,n = size(A)
     if m == 0 || n == 0
         u,s,vt = (Matrix{T}(I, m, full ? m : n), real(zeros(T,0)), Matrix{T}(I, n, n))
     else
-        u,s,vt = LAPACK.gesdd!(full ? 'A' : 'S', A)
+        if typeof(alg) == DivideAndConquer
+            u,s,vt = LAPACK.gesdd!(full ? 'A' : 'S', A)
+        elseif typeof(alg) == Simple
+            c = full ? 'A' : 'S'
+            u,s,vt = LAPACK.gesvd!(c, c, A)
+        else
+            throw(ArgumentError("Unsupported value for `alg` keyword."))
+        end
     end
     SVD(u,s,vt)
 end
 
 """
-    svd(A; full::Bool = false) -> SVD
+    svd(A; full::Bool = false, alg::SVDAlgorithm = DivideAndConquer()) -> SVD
 
 Compute the singular value decomposition (SVD) of `A` and return an `SVD` object.
 
@@ -119,6 +132,9 @@ If `full = false` (default), a "thin" SVD is returned. For a ``M
 and `V` is `N \\times N`, while in the thin factorization `U` is `M
 \\times K` and `V` is `N \\times K`, where `K = \\min(M,N)` is the
 number of singular values.
+
+If `alg = DivideAndConquer()` (default) a divide-and-conquer algorithm is used to calculate the SVD.
+One can set `alg = Simple()` to use a simple (typically slower) algorithm instead.
 
 # Examples
 ```jldoctest
@@ -144,21 +160,21 @@ julia> u == F.U && s == F.S && v == F.V
 true
 ```
 """
-function svd(A::StridedVecOrMat{T}; full::Bool = false) where T
-    svd!(copy_oftype(A, eigtype(T)), full = full)
+function svd(A::StridedVecOrMat{T}; full::Bool = false, alg::SVDAlgorithm = DivideAndConquer()) where T
+    svd!(copy_oftype(A, eigtype(T)), full = full, alg = alg)
 end
-function svd(x::Number; full::Bool = false)
+function svd(x::Number; full::Bool = false, alg::SVDAlgorithm = DivideAndConquer())
     SVD(x == 0 ? fill(one(x), 1, 1) : fill(x/abs(x), 1, 1), [abs(x)], fill(one(x), 1, 1))
 end
-function svd(x::Integer; full::Bool = false)
-    svd(float(x), full = full)
+function svd(x::Integer; full::Bool = false, alg::SVDAlgorithm = DivideAndConquer())
+    svd(float(x), full = full, alg = alg)
 end
-function svd(A::Adjoint; full::Bool = false)
-    s = svd(A.parent, full = full)
+function svd(A::Adjoint; full::Bool = false, alg::SVDAlgorithm = DivideAndConquer())
+    s = svd(A.parent, full = full, alg = alg)
     return SVD(s.Vt', s.S, s.U')
 end
-function svd(A::Transpose; full::Bool = false)
-    s = svd(A.parent, full = full)
+function svd(A::Transpose; full::Bool = false, alg::SVDAlgorithm = DivideAndConquer())
+    s = svd(A.parent, full = full, alg = alg)
     return SVD(transpose(s.Vt), s.S, transpose(s.U))
 end