+++ stdlib/LinearAlgebra/test/svd.jl
@@ -139,4 +139,35 @@ aimg  = randn(n,n)/2
     end
 end
 
+
+
+@testset "SVD Algorithms" begin
+    ≊(x,y) = isapprox(x,y,rtol=1e-15)
+
+    allpos = (v) -> begin
+        for e in v
+            e < 0 && return false
+        end
+        return true
+    end
+
+    x = [0.1 0.2; 0.3 0.4]
+
+    for alg in [LinearAlgebra.Simple(), LinearAlgebra.DivideAndConquer()]
+        sx1 = svd(x, alg = alg)
+        @test sx1.U * Diagonal(sx1.S) * sx1.Vt ≊ x
+        @test sx1.V * sx1.Vt ≊ I
+        @test sx1.U * sx1.U' ≊ I
+        @test allpos(sx1.S)
+
+        sx2 = svd!(copy(x), alg = alg)
+        @test sx2.U * Diagonal(sx2.S) * sx2.Vt ≊ x
+        @test sx2.V * sx2.Vt ≊ I
+        @test sx2.U * sx2.U' ≊ I
+        @test allpos(sx2.S)
+    end
+end
+
+
+
 end # module TestSVD